requirements.txt
tqdm
torch
torchvision
gputil
Pillow